sneaky for of [] is fine

+++ src/events/messageDelete.js
@@ -3,9 +3,7 @@ const { Event } = require('klasa');
 module.exports = class extends Event {
 
 	run(message) {
-		if (message.command && message.command.deletable && message.responses.length) {
-			for (const msg of message.responses) msg.delete();
-		}
+		if (message.command && message.command.deletable) for (const msg of message.responses) msg.delete();
 	}
 
 };
